Fix rust tests on rustc 1.82+ (#4774)
See https://releases.rs/docs/1.82.0/
Rustc used to pass the whole-archive linker flag when building tests but
no longer does.
It's now passed manually to our static libs since it's not clear how to
only do that for tests and it doesn't seem to have an impact on the size
of the rust example executable, so it's probably fine to just enable it
unconditionally.